+++ src/vegetation_tile_biomee.mod.f90
@@ -11,6 +11,7 @@ module md_vegetation_tile_biomee
   use md_cohort
   use md_cohort_linked_list
   use, intrinsic :: iso_c_binding, only: c_double
+  use, intrinsic :: ieee_arithmetic, only: ieee_is_nan
 
   ! define data types and constants
   implicit none
@@ -199,6 +200,7 @@ module md_vegetation_tile_biomee
     procedure initialize_vegn_tile
     procedure :: aggregate_pools_across_cohorts
     procedure :: lu_props
+    procedure :: export_restart_state
 
     !========= Private helper methods
 
@@ -599,14 +601,14 @@ contains
     self%n_deadtrees  = 0.0
     self%c_deadtrees  = 0.0
     self%m_turnover   = 0.0
-    self%totseed      = orgpool()
-    self%totNewC      = orgpool()
+    self%totseed      = orgpool() ! initializes with zero
+    self%totNewC      = orgpool() ! initializes with zero
 
 
     ! We reset the cohorts internal state
     it => self%cohorts()
     do while (associated(it))
-      call it%cohort%reset_cohort()
+      call it%cohort%reset_cohort_fluxes()
       it => it%next()
     end do
 
@@ -619,7 +621,7 @@ contains
     ! Reset dauly diagnostic variables
     class(vegn_tile_type), intent(inout) :: self
 
-    self%daily_fluxes = common_fluxes()
+    self%daily_fluxes = common_fluxes() ! initializes with zero
     self%dailyRh   = 0.0
     self%dailyPrcp = 0.0
     self%dailyEvap = 0.0
@@ -694,6 +696,7 @@ contains
       self%out_daily_tile(idoy, DAILY_TILE_YEAR       ) = iyears
       self%out_daily_tile(idoy, DAILY_TILE_DOY        ) = idoy
       self%out_daily_tile(idoy, DAILY_TILE_TK         ) = self%tk_daily
+      self%out_daily_tile(idoy, DAILY_TILE_TKSOIL      ) = self%tc_soil + kTkelvin
       self%out_daily_tile(idoy, DAILY_TILE_PRCP       ) = self%dailyPrcp
       self%out_daily_tile(idoy, DAILY_TILE_SOIL_W     ) = self%soilwater()
       self%out_daily_tile(idoy, DAILY_TILE_TRSP       ) = self%daily_fluxes%trsp
@@ -778,7 +781,7 @@ contains
         froot     = cc%NPProot / treeG
         fwood     = cc%NPPwood / treeG
         dDBH      = cc%dbh() - cc%DBH_ys !in m
-        dBA       = cc%basal_area() - cc%BA_ys
+        dBA       = cc%basal_area() - cc%BA_ys ! _ys: year start
 
         if (i <= NCohortMax) then
 
@@ -822,6 +825,13 @@ contains
           self%out_annual_cohorts(i, ANNUAL_COHORTS_ROOT_N     ) = cc%proot%n14
           self%out_annual_cohorts(i, ANNUAL_COHORTS_SW_N       ) = cc%psapw%n14
           self%out_annual_cohorts(i, ANNUAL_COHORTS_HW_N       ) = cc%pwood%n14
+          
+          ! print *, 'NSC,NSN', self%out_annual_cohorts(i, ANNUAL_COHORTS_NSC), self%out_annual_cohorts(i, ANNUAL_COHORTS_NSN), cc%plabl%c12, cc%plabl%n14
+          ! print *, 'Seed', self%out_annual_cohorts(i, ANNUAL_COHORTS_SEED_C     ), cc%pseed%c12, cc%pseed%n14
+          ! print *, 'Leaf', self%out_annual_cohorts(i, ANNUAL_COHORTS_LEAF_C     ), cc%pleaf%c12, cc%pleaf%n14
+          ! print *, 'Root', self%out_annual_cohorts(i, ANNUAL_COHORTS_ROOT_C     ), cc%proot%c12, cc%proot%n14
+          ! print *, 'sapw', self%out_annual_cohorts(i, ANNUAL_COHORTS_SW_C       ), cc%psapw%c12, cc%psapw%n14
+          ! print *, 'heaw', self%out_annual_cohorts(i, ANNUAL_COHORTS_HW_C       ), cc%pwood%c12, cc%pwood%n14
 
         end if
 
@@ -902,6 +912,11 @@ contains
     self%out_annual_tile(ANNUAL_TILE_DEADTREES_C     ) = 0
     self%out_annual_tile(ANNUAL_TILE_M_TURNOVER      ) = 0
     self%out_annual_tile(ANNUAL_TILE_C_TURNOVER_TIME ) = self%pwood%c12 / self%NPPW
+
+    !print *, 'MCRB_C', self%out_annual_tile(ANNUAL_TILE_MCRB_C  ), self%pmicr%c12,    self%pmicr%n14
+    !print *, 'FASTSOM', self%out_annual_tile(ANNUAL_TILE_FASTSOM ), self%psoil_fs%c12, self%psoil_fs%n14
+    !print *, 'SLOWSOM', self%out_annual_tile(ANNUAL_TILE_SLOWSOM ), self%psoil_sl%c12, self%psoil_sl%n14
+    !print *, 'INORG_N', self%out_annual_tile(ANNUAL_TILE_INORG_N ), self%inorg%n14
 
     ! Rebalance N (to compensate for the adjunction in vegn_N_uptake)
     if (inputs%params_siml%do_closedN_run) call self%recover_N_balance()
@@ -1026,12 +1041,14 @@ contains
 
     ! Local variables
     integer :: i, init_n_cohorts
+    logical :: has_restart_state
     type(cohort_type), pointer :: cc
     type(cohort_stack_item), pointer :: new
     type(params_species_biomee) :: sp
 
     ! Initialize lu_index
     self%lu_index = lu_index
+    has_restart_state = .false.
 
     ! If it is an LU acception vegetation (typically anything else than urban), we add the cohorts matching the lu_index
     if (inputs%init_lu(lu_index)%vegetated) then
@@ -1051,10 +1068,7 @@ contains
         cc%density   = inputs%init_cohort(i)%init_cohort_density ! trees/m2
         cc%age       = inputs%init_cohort(i)%init_cohort_age ! years
 
-        sp = cc%sp() ! careful this uses cc%species, ensure it is done after setting cc%species
-
-        ! set bl_max, br_max parameters ( to initialize non-structural N: plabl%n14)
-        call cc%init_bl_max_br_max()
+        sp = cc%sp() ! careful this uses cc%species, ensure cc%sp() is called only after setting cc%species
 
         ! C pools
         cc%plabl%c12 = inputs%init_cohort(i)%init_cohort_nsc
@@ -1071,10 +1085,52 @@ contains
         cc%pleaf%n14 = inputs%init_cohort(i)%init_cohort_bl_n14
         cc%proot%n14 = inputs%init_cohort(i)%init_cohort_br_n14
         cc%pseed%n14 = inputs%init_cohort(i)%init_cohort_seedC_n14
+
+        ! Other memory variables (needed for restarting a simulation)
+        if (inputs%init_cohort(i)%restart_status >= 0) then
+          cc%status = inputs%init_cohort(i)%restart_status
+          has_restart_state = .true.
+        end if
+
+        if (inputs%init_cohort(i)%restart_layer > 0) then
+          cc%layer = inputs%init_cohort(i)%restart_layer
+          has_restart_state = .true.
+        end if
+
+        if (inputs%init_cohort(i)%restart_firstlayer >= 0) then
+          cc%firstlayer = inputs%init_cohort(i)%restart_firstlayer
+          has_restart_state = .true.
+        end if
+
+        if (.not. ieee_is_nan(inputs%init_cohort(i)%restart_gdd)) then
+          cc%gdd = inputs%init_cohort(i)%restart_gdd
+          has_restart_state = .true.
+        end if
+
+        if (.not. ieee_is_nan(inputs%init_cohort(i)%restart_leaf_age)) then
+          cc%leaf_age = inputs%init_cohort(i)%restart_leaf_age
+          has_restart_state = .true.
+        end if
+
+        if (.not. ieee_is_nan(inputs%init_cohort(i)%restart_topyear)) then
+          cc%topyear = inputs%init_cohort(i)%restart_topyear
+          has_restart_state = .true.
+        end if
+
+        if (.not. ieee_is_nan(inputs%init_cohort(i)%restart_bl_max) .and. &
+            .not. ieee_is_nan(inputs%init_cohort(i)%restart_br_max)) then
+          cc%bl_max = inputs%init_cohort(i)%restart_bl_max
+          cc%br_max = inputs%init_cohort(i)%restart_br_max
+          has_restart_state = .true.
+        else
+          ! Cold starts still derive these targets from the initialized structure.
+          call cc%init_bl_max_br_max()
+        end if
+
       enddo
 
       ! Split initial layer in smaller layers (if it is full)
-      call self%relayer()
+      if (.not. has_restart_state) call self%relayer() ! restart states do not need relayer(), they already carry layer assignments
 
     end if
 
@@ -1096,10 +1152,78 @@ contains
     ! Initialize initialN0, that is used for nitrogen workaround: keep the N in the system constant at this value
     self%initialN0 = inputs%init_soil%init_N0_ecosystem
 
+    if (.not. ieee_is_nan(inputs%init_soil%restart_tk_pheno)) self%tk_pheno = inputs%init_soil%restart_tk_pheno 
+    !if (.not. ieee_is_nan(inputs%init_soil%restart_vegn_gdd)) self%gdd      = inputs%init_soil%restart_vegn_gdd
+    
     !call self%aggregate_pools_across_cohorts()
     !self%initialN0 =  self%totN
 
   end subroutine initialize_vegn_tile
+
+  subroutine export_restart_state(self, output_init_cohort, output_init_soil)
+    !////////////////////////////////////////////////////////////////////////
+    ! Export the current tile state in a format compatible with init_cohort
+    ! and init_soil, including restart-only cohort and tile fields.
+    !------------------------------------------------------------------------
+    class(vegn_tile_type), intent(in) :: self
+    real(kind=c_double), dimension(:, :), intent(out) :: output_init_cohort
+    real(kind=c_double), dimension(:), intent(out) :: output_init_soil
+
+    type(cohort_type), pointer :: cc
+    type(cohort_stack_item), pointer :: it
+    integer :: i
+
+    i = 0
+    it => self%cohorts()
+    do while (associated(it) .and. i < size(output_init_cohort, 1))
+      cc => it%cohort
+      i = i + 1
+
+      output_init_cohort(i, 1)  = cc%species
+      output_init_cohort(i, 2)  = cc%density
+      output_init_cohort(i, 3)  = cc%age
+      output_init_cohort(i, 4)  = cc%pleaf%c12
+      output_init_cohort(i, 5)  = cc%proot%c12
+      output_init_cohort(i, 6)  = cc%psapw%c12
+      output_init_cohort(i, 7)  = cc%pwood%c12
+      output_init_cohort(i, 8)  = cc%pseed%c12
+      output_init_cohort(i, 9)  = cc%plabl%c12
+      output_init_cohort(i, 10) = cc%pleaf%n14
+      output_init_cohort(i, 11) = cc%proot%n14
+      output_init_cohort(i, 12) = cc%psapw%n14
+      output_init_cohort(i, 13) = cc%pwood%n14
+      output_init_cohort(i, 14) = cc%pseed%n14
+      output_init_cohort(i, 15) = cc%plabl%n14
+      output_init_cohort(i, 16) = self%lu_index
+      output_init_cohort(i, 17) = cc%status
+      output_init_cohort(i, 18) = cc%layer
+      output_init_cohort(i, 19) = cc%firstlayer
+      output_init_cohort(i, 20) = cc%gdd
+      output_init_cohort(i, 21) = cc%leaf_age
+      output_init_cohort(i, 22) = cc%topyear
+      output_init_cohort(i, 23) = cc%bl_max
+      output_init_cohort(i, 24) = cc%br_max
+
+      it => it%next()
+    end do
+
+    output_init_soil(1)  = self%psoil_fs%c12
+    output_init_soil(2)  = self%psoil_sl%c12
+    output_init_soil(3)  = self%inorg%n14
+    output_init_soil(4)  = inputs%init_soil%N_input
+    output_init_soil(5)  = self%psoil_fs%n14
+    output_init_soil(6)  = self%psoil_sl%n14
+    output_init_soil(7)  = self%pmicr%c12
+    output_init_soil(8)  = self%pmicr%d13
+    output_init_soil(9)  = self%pmicr%n14
+    output_init_soil(10) = self%wcl(1)
+    output_init_soil(11) = self%wcl(2)
+    output_init_soil(12) = self%wcl(3)
+    output_init_soil(13) = self%initialN0
+    output_init_soil(14) = self%tk_pheno
+    !output_init_soil(15) = self%gdd ! previously we had a vegn%gdd, but use only cc%gdd
+
+  end subroutine export_restart_state
 
   !----------------------------------------------------------------
   ! Private helper methods